requirements-ml-dev.txt
# ML Development Dependencies
# For development machine (training)
# Install with: pip install -r requirements-ml-dev.txt

# PyTorch (install separately based on CUDA version)
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CPU-only: pip install torch torchvision torchaudio

# TorchSig (RF signal datasets)
torchsig

# Tests
pytest>=7.0.0

# Training dependencies
numpy>=1.23,<2.0
scikit-learn>=1.0.0
tensorboard>=2.10.0

# ONNX export
onnx>=1.14.0
onnxruntime>=1.16.0  # For testing ONNX export locally

# Data handling and utilities
matplotlib>=3.5.0
tqdm>=4.64.0
pandas>=1.5.0
pillow>=9.0.0

# Base SPEAR-Edge dependencies (for testing with captures)
fastapi>=0.100,<0.115
uvicorn>=0.23
pydantic>=2.0
python-dotenv
websockets>=11
requests>=2.28.0
